+++ scripts/less.js
@@ -4,10 +4,16 @@
 const less  = require('less');
 const fs    = require('fs');
 const path  = require('path');
+const chalk = require('chalk');
 
-console.log(`Starting LESS compiler ...`);
+let err     = chalk.bold.red;
+let warn    = chalk.yellow;
+let info    = chalk.keyword('cornflowerblue');
+let success = chalk.bold.green;
 
-console.info(`Current working directory: ${__dirname}`);
+console.log(info(`Starting LESS compiler ...`));
+
+console.info(info(`Current working directory: ${__dirname}`));
 
 // If you wanted to render a file, you would first read it into a string
 // (to pass to less.render) and then set the filename field on options to be
@@ -20,22 +26,22 @@ let themeTarget = path.join(__dirname, '../source/renderer/assets/css/theme-zett
 let mainLess    = '';
 let themeLess   = '';
 
-console.log(`Reading input files ...`);
+console.log(info(`Reading input files ...`));
 try {
     mainLess = fs.readFileSync(mainFile, { encoding: 'utf8' });
-    console.log(`Successfully read ${mainFile}!`);
+    console.log(success(`Successfully read ${mainFile}!`));
 }catch(e) {
-    console.error(`ERROR: Could not read ${mainFile}: ${e.name}`);
-    console.error(e.message);
+    console.error(err(`ERROR: Could not read ${mainFile}: ${e.name}`));
+    console.error(err(e.message));
     return;
 }
 
 try {
     themeLess = fs.readFileSync(themeFile, { encoding: 'utf8' });
-    console.log(`Successfully read ${themeFile}`);
+    console.log(success(`Successfully read ${themeFile}`));
 }catch(e) {
-    console.error(`ERROR: Could not read ${themeFile}: ${e.name}`);
-    console.error(e.message);
+    console.error(err(`ERROR: Could not read ${themeFile}: ${e.name}`));
+    console.error(err(e.message));
     return;
 }
 
@@ -43,26 +49,26 @@ try {
  * PART ONE
  */
 
-console.log(`Compiling ${mainFile} ...`);
+console.log(info(`Compiling ${mainFile} ...`));
 less.render(mainLess, {
     'filename': mainFile
 })
 .then(function(output) {
-    console.log(`Done compiling ${mainFile}! Writing to target file ...`);
+    console.log(success(`Done compiling ${mainFile}! Writing to target file ...`));
     try {
         fs.writeFileSync(mainTarget, output.css, { encoding: 'utf8' });
-        console.log(`Done writing CSS to file ${mainTarget}!`);
-        console.info(`Sourcemap:`, output.map);
-        console.info(`Imported files:`, output.imports);
+        console.log(success(`Done writing CSS to file ${mainTarget}!`));
+        console.info(info(`Sourcemap:`, output.map));
+        console.info(info(`Imported files: [\n   `, output.imports.join(',\n    '), '\n]'));
     } catch(e) {
-        console.error(`ERROR: Error on writing ${mainFile}: ${e.name}`);
-        console.error(e.message);
+        console.error(err(`ERROR: Error on writing ${mainFile}: ${e.name}`));
+        console.error(err(e.message));
     }
 },
 function(error) {
     if(error) {
-        console.error(`ERROR: Could not compile ${mainLess}: ${error.name}`);
-        console.error(error.message);
+        console.error(err(`ERROR: Could not compile ${mainLess}: ${error.name}`));
+        console.error(err(error.message));
     }
 });
 
@@ -70,25 +76,25 @@ function(error) {
  * PART TWO
  */
 
-console.log(`Compiling ${themeFile} ...`);
+console.log(info(`Compiling ${themeFile} ...`));
 less.render(themeLess, {
     'filename': themeFile
 })
 .then(function(output) {
-    console.log(`Done compiling ${themeFile}! Writing to target file ...`);
+    console.log(success(`Done compiling ${themeFile}! Writing to target file ...`));
     try {
         fs.writeFileSync(themeTarget, output.css, { encoding: 'utf8' });
-        console.log(`Done writing CSS to file ${themeTarget}!`);
-        console.info(`Sourcemap:`, output.map);
-        console.info(`Imported files:`, output.imports);
+        console.log(success(`Done writing CSS to file ${themeTarget}!`));
+        console.info(info(`Sourcemap:`, output.map));
+        console.info(info(`Imported files: [\n   `, output.imports.join(',\n    '), '\n]'));
     } catch(e) {
-        console.error(`ERROR: Error on writing ${themeFile}: ${e.name}`);
-        console.error(e.message);
+        console.error(err(`ERROR: Error on writing ${themeFile}: ${e.name}`));
+        console.error(err(e.message));
     }
 },
 function(error) {
     if(error) {
-        console.error(`ERROR: Could not compile ${themeLess}: ${error.name}`);
-        console.error(error.message);
+        console.error(err(`ERROR: Could not compile ${themeLess}: ${error.name}`));
+        console.error(err(error.message));
     }
 });